requirements.txt
pyTelegramBotAPI
schedule
requests
Pillow